feat: Update dependencies and add flutter_heatmap_calendar package
This commit updates the dependencies in the pubspec.yaml file to include the latest stable versions of webview_flutter, webview_flutter_android, and webview_flutter_wkwebview. Additionally, it adds the flutter_heatmap_calendar package with version 1.0.5.

The flutter_heatmap_calendar package provides functionality for displaying a heatmap calendar in the application, which can be used to visualize data related to dates. This package will be utilized in future development to enhance the user experience and provide valuable insights.

Please note that this commit also includes changes to the devtools_options.yaml and pubspec.lock files to reflect the addition of the flutter_heatmap_calendar package.